% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjMat.R
\docType{methods}
\name{adjMat}
\alias{adjMat}
\alias{adjMat,miRmapper-method}
\title{adjMat}
\usage{
adjMat(object)

\S4method{adjMat}{miRmapper}(object)
}
\arguments{
\item{object}{a 'miRmapper' object}
}
\value{
a matrix object of the interaction network
}
\description{
This generic is for use on the miRmapper class to generate an
adjacency matrix of the interaction network. The matrix has a column
for each miRNA and a row for each gene. 0 indicates no interaction,
while a 1 indicates an interaction.
}
\examples{
adjMat(Object)

}
