% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barPlot.R
\docType{methods}
\name{barPlot}
\alias{barPlot}
\alias{barPlot,miRmapper-method}
\title{barPlot}
\usage{
barPlot(object)

\S4method{barPlot}{miRmapper}(object)
}
\arguments{
\item{object}{a 'miRmapper' object}
}
\value{
plot bar graph of the impact percentages for the miRNAs
}
\description{
This generic is for use on the 'miRmapper' class, used to generate a
bar plot that depicts the impact of each miRNA of the set of gene targets.
}
\examples{
barPlot(Object)

}
