% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrogram.R
\docType{methods}
\name{dendrogram}
\alias{dendrogram}
\alias{dendrogram,miRmapper-method}
\title{dendrogram}
\usage{
dendrogram(object)

\S4method{dendrogram}{miRmapper}(object)
}
\arguments{
\item{object}{a 'miRmapper' object}
}
\value{
plot dendrogram figure of miRNAs
}
\description{
This generic is for use on the miRmapper class, used to generate a
dendrogram for miRNA clustering based on shared targets
}
\examples{
dendrogram(Object)

}
