% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAdjmat.R
\docType{methods}
\name{getAdjmat}
\alias{getAdjmat}
\alias{getAdjmat,miRmapper-method}
\title{getAdjmat}
\usage{
getAdjmat(object)

\S4method{getAdjmat}{miRmapper}(object)
}
\arguments{
\item{object}{a 'miRmapper' object}
}
\value{
Adjacency matrix of the interactions
}
\description{
This generic is for use on the 'miRmapper' class that returns the
adjacency matrix generated from the interactions.
}
\examples{
getAdjmat(Object)

}
