% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDEgenes.R
\docType{methods}
\name{getDEgenes}
\alias{getDEgenes}
\alias{getDEgenes,miRmapper-method}
\title{getDEgenes}
\usage{
getDEgenes(object)

\S4method{getDEgenes}{miRmapper}(object)
}
\arguments{
\item{object}{a 'miRmapper' object}
}
\value{
a list of differentiall expressed genes
}
\description{
This generic is for use on the miRmapper class that returns the 'DEgenes'
slot of the 'miRmapper' object.
}
\examples{
getDEgenes(Object)

}
