% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getImpact.R
\docType{methods}
\name{getImpact}
\alias{getImpact}
\alias{getImpact,miRmapper-method}
\title{getImpact}
\usage{
getImpact(object)

\S4method{getImpact}{miRmapper}(object)
}
\arguments{
\item{object}{a 'miRmapper' object}
}
\value{
a dataframe object of impact metrics for each miRNA
}
\description{
This generic is for use on the miRmapper class to generate metrics that
measure the predicted impact each miRNA has on the differentially expression genes.
}
\examples{
getImpact(Object)

}
