% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInteractions.R
\docType{methods}
\name{getInteractions}
\alias{getInteractions}
\alias{getInteractions,miRmapper-method}
\title{getInteractions}
\usage{
getInteractions(object)

\S4method{getInteractions}{miRmapper}(object)
}
\arguments{
\item{object}{a 'miRmapper' object}
}
\value{
a dataframe of miRNA to gene interactions
}
\description{
This generic is for use on the miRmapper class that returns the
'interactions' slot of a miRmapper object.
}
\examples{
getInteractions(Object)

}
