% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identityPlot.R
\docType{methods}
\name{identityPlot}
\alias{identityPlot}
\alias{identityPlot,miRmapper-method}
\title{identityPlot}
\usage{
identityPlot(object)

\S4method{identityPlot}{miRmapper}(object)
}
\arguments{
\item{object}{a 'miRmapper' object}
}
\value{
plot a heatmap of the miRNAs
}
\description{
This generic is for use on the miRmapper class, used to generate a
heatmap representing an identity plot that represents the similarity
between miRNAS. A darker colors indicate a cluster where interactions are
shared.
}
\examples{
identityPlot(Object)

}
