% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miRmapper.R
\name{miRmapper-class}
\alias{miRmapper-class}
\alias{miRmapper}
\title{An S4 class to represent miRNA to gene interactions}
\usage{
miRmapper(interations = list(), DEgenes = list())
}
\description{
An S4 class to represent miRNA to gene interactions
}
\section{Slots}{

\describe{
\item{\code{interactions}}{a list of miRNA to gene interactions (2 columns)}

\item{\code{DEgenes}}{a list of differentially expressed genes from a sequencing experiment}

\item{\code{adjmat}}{a matrix object generated from the 'setAdjMat' method}
}}

