% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAnalysis.R
\docType{methods}
\name{runAnalysis}
\alias{runAnalysis}
\alias{runAnalysis,miRmapper-method}
\title{runAnalysis}
\usage{
runAnalysis(object)

\S4method{runAnalysis}{miRmapper}(object)
}
\arguments{
\item{object}{a 'miRmapper' object}
}
\value{
5 output files into a directory named "miRmapper-Output". The files
consist of a table indicating the total predicted gene targets for each miRNA
and its percentage of impact, a table of the adjacency matrix that details
the miRNA-target network, a bar plot of the the miRNA impact percentages, a
dedrogram depicting miRNA clusters, and an identity heatmap depicting miRNA
clustering.
}
\description{
This generic is for use on the miRmapper class to perform the the entire
miRmapper analysis and export all outputs into a directory named
'miRmapper-Output' that is placed in the useres working directory.
}
\examples{
runAnalysis(Object)

}
