% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setDEgenes.R
\docType{methods}
\name{setDEgenes}
\alias{setDEgenes}
\alias{setDEgenes,ANY-method}
\alias{setDEgenes,miRmapper-method}
\title{setDEgenes}
\usage{
setDEgenes(object, genes)

\S4method{setDEgenes}{ANY}(object, genes)
}
\arguments{
\item{object}{a 'miRmapper' object}

\item{genes}{a 'list' object containing one column of gene symbols}
}
\value{
None
}
\description{
This generic is for use on the miRmapper class, used to set the 'DEgenes'
slot of a 'miRmapper' object.
}
\examples{
setDEgenes(Object, genes)

}
