% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setInteractions.R
\docType{methods}
\name{setInteractions}
\alias{setInteractions}
\alias{setInteractions,ANY-method}
\alias{setInteractions,miRmapper-method}
\title{setInteractions}
\usage{
setInteractions(object, interactions.list)

\S4method{setInteractions}{ANY}(object, interactions.list)
}
\arguments{
\item{object}{a 'miRmapper' object}

\item{interactions.list}{a 'list' object containing two columns with each row
representing a miRNA to gene interaction.}
}
\value{
None
}
\description{
This generic is for use on the miRmapper class, used to set the
'interactions' slot of a 'miRmapper' object.
}
\examples{
setInteractions(Object, interactions.list)

}
